+++ app/assets/stylesheets/05-page.scss
@@ -5,7 +5,7 @@
     flex-direction: column;
     align-items: flex-start;
     background-color: white;
-    background-image: url("https://versify-dev.s3.amazonaws.com/images/ui/universe-background.jpeg");
+    background-image: url("https://versify-dev.s3.amazonaws.com/images/ux/universe-background.jpeg");
     background-size: cover;
     background-repeat: repeat;
     overflow-y: hidden;
