+++ frontend/reducers/root_reducer.js
@@ -1,16 +1,14 @@
-import { combineReducers } from 'redux';
-import entitiesReducer from './entities_reducer';
-import uiReducer from './ui_reducer';
-import sessionReducer from './session_reducer';
-import errorsReducer from './errors_reducer';
+import { combineReducers } from "redux";
+import entitiesReducer from "./entities_reducer";
+import uxReducer from "./ux_reducer";
+import sessionReducer from "./session_reducer";
+import errorsReducer from "./errors_reducer";
 
 const rootReducer = combineReducers({
-    entities: entitiesReducer,
-    ui: uiReducer,
-    session: sessionReducer,
-    errors: errorsReducer,
-
-})
+	entities: entitiesReducer,
+	ux: uxReducer,
+	session: sessionReducer,
+	errors: errorsReducer,
+});
 
 export default rootReducer;
-
