--- frontend/reducers/ui_reducers/playlist_nav_dropdown_reducer.js
@@ -1,17 +1,0 @@
-import { PLAYLISTNAV_DROPDOWN_OPEN,
-    PLAYLISTNAV_DROPDOWN_CLOSE,
-} from "../../actions/ui_actions";
-
-const playlistNavDropdownReducer = (playlistNavDropdownState = {isOpen: false}, action) => {
-    Object.freeze(playlistNavDropdownState);
-    switch (action.type) {
-        case PLAYLISTNAV_DROPDOWN_OPEN:
-            return {isOpen: true};
-        case PLAYLISTNAV_DROPDOWN_CLOSE:
-            return {isOpen: false};
-        default:
-            return playlistNavDropdownState;
-    }
-}
-
-export default playlistNavDropdownReducer;
